requirements.txt
nose
sphinx
MySQL-python